# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Playbasis API",
        "description": "API Explorer",
        "termsOfService": "",
        "contact": {
            "email": "info@playbasis.com"
        },
        "version": "1.0"
    },
    "host": "api.pbapp.net",
    "basePath": "/",
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/Push/send": {
            "post": {
                "tags": [
                    "Communication"
                ],
                "description": "Send push notfications to a player",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token returned by Playbasis Authentication",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "description": "SMS Message",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "template_id",
                        "in": "query",
                        "description": "Template message",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Push/goods": {
            "post": {
                "tags": [
                    "Communication"
                ],
                "description": "Send coupon to a player via push notification.",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token returned by Playbasis Authentication",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "ref_id",
                        "in": "query",
                        "description": "Reference transaction ID for redemption",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "description": "SMS Message | use {{coupon}} for the code | message or template_id is required",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "template_id",
                        "in": "query",
                        "description": "Template message | message or template_id is required",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Push/deviceRegistration": {
            "post": {
                "tags": [
                    "Communication"
                ],
                "description": "Register for your device",
                "parameters": [
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token returned by Playbasis Authentication",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "device_token",
                        "in": "query",
                        "description": "Access token returned from device",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "device_description",
                        "in": "query",
                        "description": "Device model description",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "device_name",
                        "in": "query",
                        "description": "Device model name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "os_type",
                        "in": "query",
                        "description": "Device model name",
                        "required": true,
                        "type": "string",
                        "default": "IOS",
                        "enum": [
                            "IOS",
                            "Android"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Push/deviceDeRegistration": {
            "post": {
                "tags": [
                    "Communication"
                ],
                "description": "DeRegister for your device",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token returned by Playbasis Authentication",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "device_token",
                        "in": "query",
                        "description": "Access token returned from device",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Push/recent": {
            "get": {
                "tags": [
                    "Communication"
                ],
                "description": "List recent push notifications to a player",
                "parameters": [
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "since",
                        "in": "query",
                        "description": "DateTime format string",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Push/template/{template_id}": {
            "get": {
                "tags": [
                    "Communication"
                ],
                "description": "List recent push notifications to a player",
                "parameters": [
                    {
                        "name": "template_id",
                        "in": "path",
                        "description": "Template message",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "apiKey": {
            "type": "apiKey",
            "name": "api_key",
            "in": "query"
        },
        "token": {
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    },
    "security": [
        {
            "apiKey": [],
            "token": []
        }
    ]
}